iris_neuralnet/requirements.txt
pandas==0.25.1
matplotlib==3.1.1
scikit-learn==0.21.3
djangorestframework==3.14.0
